requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
django-cors-headers
